Money and save manager
Money counter done, money ref the global system found in the savemanager. the money counter is located in the player movements. moneyTxt needs to be set up in the inspector. the money needs a script for when used in a shop and decreases the money amount globally and updates the money counter.